requirements.txt
fastapi[all]
uvicorn
python-dotenv
webauthn
itsdangerous